Boarding-pass landing + robust copy-invite
- Replace the plain React landing with a static boarding-pass HTML at
  public/landing.html, served at "/" by Vite (configureServer middleware)
  and Express (prod). A small injected <script> wires the two tabs (new
  departure / i have a ticket) and the BOARD button to /api/groups +
  /g/:id navigation.
- Fix copyInvite() so it works in non-secure contexts (LAN HTTP): try
  navigator.clipboard, fall back to a hidden textarea + execCommand, and
  as a last resort surface the URL via window.prompt so the user can
  copy by hand.
- Move Express's listen port from PORT to API_PORT so it doesn't collide
  with dev tooling that sets PORT for Vite.
- Includes prior in-progress drag-to-reorder migration + handler changes.

+++ api/groups/[id].js
@@ -12,7 +12,7 @@ export default async function handler(req, res) {
 
   const rows = await sql`
     SELECT g.id, g.name, g.created_at,
-           COUNT(m.id)::int AS member_count
+           COUNT(m.id) AS member_count
       FROM groups g
       LEFT JOIN members m ON m.group_id = g.id
      WHERE g.id = ${id}
